AppComponent doesn't handle the logis of a position as it's moved to a separate component
