+++ src/app/app-routing.module.ts
@@ -1,10 +1,15 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { AppComponent } from './app.component';
+import { PositionsTableComponent } from './positions-table/positions-table.component';
 
-const routes: Routes = [];
+const routes: Routes = [
+  { path: 'positions', component: PositionsTableComponent },
+  // { path: '', redirectTo: '/positions', pathMatch: 'full' }, // Redirect to 'positions' by default
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule]
+  exports: [RouterModule],
 })
-export class AppRoutingModule { }
+export class AppRoutingModule {}
