+++ src/app/app.component.html
@@ -1,123 +1,3 @@
-<nav class="navbar navbar-expand-lg bg-body-tertiary">
-  <div class="container-fluid">
-    <a class="navbar-brand" href="#">IBKR Dashboard</a>
-    <button
-      class="navbar-toggler"
-      type="button"
-      data-bs-toggle="collapse"
-      data-bs-target="#navbarSupportedContent"
-      aria-controls="navbarSupportedContent"
-      aria-expanded="false"
-      aria-label="Toggle navigation"
-    >
-      <span class="navbar-toggler-icon"></span>
-    </button>
-    <div class="collapse navbar-collapse" id="navbarSupportedContent">
-      <ul class="navbar-nav me-auto mb-2 mb-lg-0">
-        <li class="nav-item">
-          <a class="nav-link active" aria-current="page" href="#">Home</a>
-        </li>
-        <li class="nav-item dropdown">
-          <a
-            class="nav-link dropdown-toggle"
-            href="#"
-            role="button"
-            data-bs-toggle="dropdown"
-            aria-expanded="false"
-          >
-            Socials
-          </a>
-          <ul class="dropdown-menu">
-            <li>
-              <a
-                class="dropdown-item"
-                href="https://github.com/EconLQ"
-                target="_blank"
-                >GitHub</a
-              >
-            </li>
-            <li>
-              <a
-                class="dropdown-item"
-                href="https://twitter.com/zavbog_LQ"
-                target="_blank"
-                >Twitter</a
-              >
-            </li>
-            <li><hr class="dropdown-divider" /></li>
-            <li><a class="dropdown-item" href="#">Something else here</a></li>
-          </ul>
-        </li>
-      </ul>
-      <form class="d-flex" role="search">
-        <input
-          class="form-control me-2"
-          type="search"
-          (ngModelChange)="searchPositions(key.value)"
-          #key="ngModel"
-          ngModel
-          name="key"
-          placeholder="Search positions..."
-          aria-label="Search"
-        />
-      </form>
-    </div>
-  </div>
-</nav>
-<h3 id="position-title">Positions Table</h3>
-<p id="position-subtitle">
-  Includes all of the positons from your portfolio that you've recently took
-</p>
-<div id="main-container">
-  <table
-    class="table table-dark table-stripped-columns table-bordered table-hover"
-    id="positions-table"
-  >
-    <thead>
-      <tr>
-        <th scope="col">Contract Id</th>
-        <th scope="col">Ticker</th>
-        <th scope="col">Entry date</th>
-        <th scope="col">Position Size</th>
-        <th scope="col">Unrealized PnL</th>
-        <th scope="col">Realized PnL</th>
-        <th scope="col">Average Cost</th>
-        <th scope="col">Last Price</th>
-        <th scope="col"></th>
-        <th scope="col"></th>
-      </tr>
-    </thead>
-    <tbody>
-      <tr *ngFor="let position of positions">
-        <td>{{ position.contractId }}</td>
-        <td>{{ position.ticker }}</td>
-        <td>{{ position.date }}</td>
-        <td>{{ position.position }}</td>
-        <td>{{ position.unrealizedPnL }}</td>
-        <td>{{ position.realizedPnL }}</td>
-        <td>{{ position.averageCost }}</td>
-        <td>{{ position.lastMarketPrice }}</td>
-        <td>
-          <button
-            type="button"
-            class="btn btn-primary"
-            data-bs-toggle="modal"
-            (click)="onOpenModal(position, 'edit')"
-          >
-            Edit
-          </button>
-        </td>
-        <td>
-          <button
-            type="button"
-            class="btn btn-danger"
-            data-bs-toggle="modal"
-            (click)="onOpenModal(position, 'delete')"
-          >
-            Delete
-          </button>
-        </td>
-      </tr>
-    </tbody>
-  </table>
-</div>
+<!-- injecting navbar -->
+<app-navbar></app-navbar>
+<router-outlet></router-outlet>
