+++ src/app/delete-position-modal/delete-position-modal.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from '@angular/core';
-import { Position } from '../position';
+import { Position } from '../positions-table/position';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
-import { PositionService } from '../position.service';
+import { PositionService } from '../positions-table/position.service';
 import { HttpErrorResponse } from '@angular/common/http';
 
 @Component({
